// foundation-layer/src/platform-access/enums/action-key.enum.ts
export enum ActionKey {
  VIEW_EDITORIAL_QUEUE = "VIEW_EDITORIAL_QUEUE",
  REVIEW_CANDIDATE = "REVIEW_CANDIDATE",
  APPROVE_PUBLICATION = "APPROVE_PUBLICATION",
  REJECT_PUBLICATION = "REJECT_PUBLICATION",
  SCHEDULE_HANDOFF = "SCHEDULE_HANDOFF",
  VIEW_RELIABILITY_REPORTS = "VIEW_RELIABILITY_REPORTS",
  TRIGGER_RELIABILITY_GATE = "TRIGGER_RELIABILITY_GATE",
  VIEW_LEARNING_INSIGHTS = "VIEW_LEARNING_INSIGHTS",
  MANAGE_ROLE_ASSIGNMENTS = "MANAGE_ROLE_ASSIGNMENTS",
  MANAGE_PERMISSION_POLICIES = "MANAGE_PERMISSION_POLICIES",
  MANAGE_WORKSPACES = "MANAGE_WORKSPACES",
  VIEW_PLATFORM_AUDIT = "VIEW_PLATFORM_AUDIT",
  OVERRIDE_EDITORIAL_DECISION = "OVERRIDE_EDITORIAL_DECISION",
  OVERRIDE_RELIABILITY_GATE = "OVERRIDE_RELIABILITY_GATE",
  PUBLISH_ARTIFACT = "PUBLISH_ARTIFACT",
  VIEW_PUBLICATION_PACKAGE = "VIEW_PUBLICATION_PACKAGE",
}